fix(dosage): corrige le compteur journalier de sécurité qui restait à 0
dailyPhInjectedMl et dailyOrpInjectedMl étaient unsigned long, donc
chaque petit incrément (~0.17 ml/cycle) était tronqué à 0 par le cast.
La limite journalière n'était ainsi jamais atteinte malgré des volumes
dépassant le seuil configuré.

Correction : types passés en float, cast supprimé dans updateSafetyTracking.

+++ src/config.h
@@ -137,8 +137,8 @@ struct ProductConfig {
 struct SafetyLimits {
   float maxPhMinusMlPerDay = 500.0f;
   float maxChlorineMlPerDay = 300.0f;
-  unsigned long dailyPhInjectedMl = 0;
-  unsigned long dailyOrpInjectedMl = 0;
+  float dailyPhInjectedMl = 0.0f;
+  float dailyOrpInjectedMl = 0.0f;
   unsigned long dayStartTimestamp = 0;
   bool phLimitReached = false;
   bool orpLimitReached = false;
